working with file at once in assm, count commands
assm also counts the amount of commands (and arguments, in other words, the amount of numbers in byte file) and writes it to file. I use a lib from my onegin repository for working with source code file. assm also makes a binary file (but now it's name is hard coded)